refactor: AuthModal을 template으로 분리
