+++ src/components/GlobalNavigationBar/GlobalNavigationBar.tsx
@@ -7,11 +7,9 @@ import Image from 'next/image';
 import Link from 'next/link';
 import { usePathname, useRouter } from 'next/navigation';
 
-import userApi from '@/apis/user/user';
-import AuthModal from '@/features/auth/components/AuthModal/AuthModal';
-import { useAuthActions, useIsOpenSignUpModal } from '@/features/auth/store';
+import { useAuthActions } from '@/features/auth/store';
 import { ROUTES } from '@/shared/constants/routes';
-import { useUserImageUrl, useUserNickname } from '@/shared/store/user';
+import { useUserImageUrl } from '@/shared/store/user';
 import { tw } from '@/shared/utils/tailwindMerge';
 
 import Button from '../Button/Button';
@@ -34,25 +32,13 @@ type GlobalNavigationBarProps = ComponentPropsWithoutRef<'header'> & {
 const GlobalNavigationBar = ({ className, isSignedIn, isRequesting, ...props }: GlobalNavigationBarProps) => {
   const rootClassName = tw(styles.root, className);
   const pathName = usePathname();
-  const isOpenSignUpModal = useIsOpenSignUpModal();
-  const nickname = useUserNickname();
   const profileImgUrl = useUserImageUrl();
-  const { setIsOpenSignUpModal, setIsSignedIn } = useAuthActions();
+  const { setIsOpenSignUpModal } = useAuthActions();
   const router = useRouter();
 
   const handleClickLoginButton = () => {
     setIsOpenSignUpModal(true);
     router.push('/?steps=signUp');
-  };
-
-  const handleClickCloseButton = () => {
-    setIsOpenSignUpModal(false);
-    router.replace('/');
-  };
-
-  const handleAbortSignUp = async () => {
-    await userApi.patch({ nickname, field: null });
-    setIsSignedIn(true);
   };
 
   return (
@@ -114,7 +100,6 @@ const GlobalNavigationBar = ({ className, isSignedIn, isRequesting, ...props }: 
           </Button>
         )}
       </header>
-      <AuthModal isOpen={isOpenSignUpModal} onClose={handleClickCloseButton} onAbortSignUp={handleAbortSignUp} />
     </>
   );
 };
